+++ features/overlays/seaCreaturesHpTracker.js
@@ -1,17 +1,19 @@
 import settings, { allOverlaysGui, seaCreaturesHpOverlayGui } from "../../settings";
-import { AQUA, BOLD } from "../../constants/formatting";
+import { AQUA, BOLD, RED } from "../../constants/formatting";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { getWorldName, isInSkyblock } from "../../utils/playerState";
 import { BACKWATER_BAYOU, CRIMSON_ISLE, CRYSTAL_HOLLOWS, JERRY_WORKSHOP, WATER_FISHING_WORLDS, WATER_HOTSPOT_WORLDS } from "../../constants/areas";
 import { OFF_SOUND_MODE } from "../../constants/sounds";
 import { registerIf } from "../../utils/registers";
 import { getSeaCreaturesInRange } from "../../utils/entityDetection";
+import { getMcEntityById } from "../../utils/common";
 
 const LOOTSHARE_DISTANCE = 30;
 const TRACKED_MOBS = [
     {
         worlds: [CRIMSON_ISLE],
         baseMobName: 'Fiery Scuttler',
+        hasImmunity: true,
     },
     {
         worlds: [CRIMSON_ISLE],
@@ -20,6 +22,7 @@ const TRACKED_MOBS = [
     {
         worlds: [CRIMSON_ISLE],
         baseMobName: 'Thunder',
+        hasImmunity: true,
     },
     {
         worlds: [CRIMSON_ISLE],
@@ -36,14 +39,17 @@ const TRACKED_MOBS = [
     {
         worlds: [JERRY_WORKSHOP],
         baseMobName: 'Yeti',
+        hasImmunity: true,
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
         baseMobName: 'Alligator',
+        hasImmunity: true,
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
         baseMobName: 'Blue Ringed Octopus',
+        hasImmunity: true,
     },
     {
         worlds: WATER_HOTSPOT_WORLDS,
@@ -56,32 +62,46 @@ const TRACKED_MOBS = [
     {
         worlds: [BACKWATER_BAYOU],
         baseMobName: 'Titanoboa',
+        hasImmunity: true,
     },
     {
         worlds: [CRYSTAL_HOLLOWS],
         baseMobName: 'Abyssal Miner',
+        hasImmunity: true,
     },
     {
         worlds: WATER_FISHING_WORLDS,
         baseMobName: 'The Sea Emperor',
+        hasImmunity: true,
     },
     {
         worlds: WATER_FISHING_WORLDS,
         baseMobName: 'Water Hydra',
+        hasImmunity: true,
     },
     {
         worlds: WATER_FISHING_WORLDS,
         baseMobName: 'Phantom Fisher',
+        hasImmunity: true,
     },
     {
         worlds: WATER_FISHING_WORLDS,
         baseMobName: 'Grim Reaper',
+        hasImmunity: true,
     },
     {
         worlds: WATER_FISHING_WORLDS,
         baseMobName: 'Great White Shark',
+        hasImmunity: true,
+    },
+    {
+        worlds: WATER_FISHING_WORLDS,
+        baseMobName: 'Squid',
+        hasImmunity: true,
     },
 ];
+
+const TRACKED_MOB_NAMES = TRACKED_MOBS.map(n => n.baseMobName);
 
 const TRACKED_WORLD_NAMES = TRACKED_MOBS
     .map(function (m) { return m.worlds; })
@@ -104,6 +124,7 @@ register("worldUnload", () => {
     mobs = [];
 });
 
+// Track seen entity ID and based on it define ticksexisted
 function trackSeaCreaturesHp() {
     try {
         if (!settings.seaCreaturesHpOverlay ||
@@ -113,10 +134,20 @@ function trackSeaCreaturesHp() {
             return;
         }
     
-        const currentMobs = getSeaCreaturesInRange(TRACKED_MOBS.map(n => n.baseMobName), LOOTSHARE_DISTANCE)
+        const currentMobs = getSeaCreaturesInRange(TRACKED_MOB_NAMES, LOOTSHARE_DISTANCE)
             .sort((a, b) => a.currentHpNumber - b.currentHpNumber) // Lowest HP comes first
-            .map(sc => ({ nametag: sc.shortNametag, baseMobName: sc.baseMobName }))
-            .slice(0, settings.seaCreaturesHpOverlay_maxCount); // Top N
+            .slice(0, settings.seaCreaturesHpOverlay_maxCount) // Top N
+            .map(sc => {
+                const mobEntity = getMcEntityById(sc.mcEntityId - 1);
+                const ticksExisted = mobEntity && mobEntity instanceof net.minecraft.entity.Entity
+                    ? mobEntity.field_70173_aa
+                    : 0;
+                console.log((sc.mcEntityId - 1).toString() + ' existed for Ticks ' + ticksExisted);
+                const isImmune = TRACKED_MOBS.find(m => m.baseMobName === sc.baseMobName)?.hasImmunity 
+                    ? ticksExisted < 20 * 5 // field_70173_aa -> ticksExisted less than 5 seconds
+                    : false;
+                return { nametag: sc.shortNametag, baseMobName: sc.baseMobName, isImmune: isImmune };
+            }); 
 
         const addedMobNames = currentMobs.filter(cm => {
             return !mobs.some(m => m.baseMobName === cm.baseMobName);
@@ -125,7 +156,7 @@ function trackSeaCreaturesHp() {
         if (
             currentMobs.length > mobs.length &&
             settings.soundMode !== OFF_SOUND_MODE &&
-            !addedMobNames.every(m => m.baseMobName === 'Reindrake') // Reindrake flies around and goes out of nametags render distance periodically, we don't need sound for it
+            !addedMobNames.every(m => m.baseMobName === 'Reindrake') // Reindrake flies around and goes out of nametags render distance periodically, we don't need sound when detecting it
         ) {
             World.playSound('random.orb', 0.75, 1);
         }
@@ -155,7 +186,8 @@ function renderHpOverlay() {
     if (mobs.length) {
         let overlayText = `${AQUA}${BOLD}Sea creatures HP\n`;
         mobs.forEach((mob) => {
-            overlayText += `${mob.nametag}\n`;
+            const immunityText = mob.isImmune ? ` ${RED}${BOLD}[Immune]` : '';
+            overlayText += `${mob.nametag}${immunityText}\n`;
         });
         drawText(overlayText);
     }