+++ settings.js
@@ -1063,7 +1063,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Overlays",
     configName: "seaCreaturesHpOverlay",
     title: "Sea creatures HP",
-    description: `Shows an overlay with the HP of nearby Legendary/Mythic sea creatures when they're in lootshare range. Sea creatures with lower HP come first.`,
+    description: `Shows an overlay with the HP of nearby Legendary/Mythic sea creatures when they're in lootshare range. It also has immunity indicator for sea creatures that have 5 seconds damage reduction period.`,
     subcategory: "Sea creatures HP",
     value: true
 })
@@ -1071,7 +1071,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Overlays",
     configName: "seaCreaturesHpOverlay_maxCount",
     title: "Maximum sea creatures HP count",
-    description: "Show maximum N sea creatures nearby (to limit overlay size).",
+    description: "Show maximum N sea creatures nearby (to limit overlay size). Sea creatures with lower HP come first.",
     options: [1, 20],
     value: 5,
     subcategory: "Sea creatures HP"